Fix virtual headers symlink action for C++ public headers
The virtual headers action should be executed when strip_include_prefix or include_prefix
attribute is not 'None', because the "." attribute values are resolved to empty string,
which is a valid value.

The virtual haeders action should be independent from the original header path and
include path to support library target definitions at any BUILD file with any project layout.